Modularize the WASM VM code in preparation for upstreaming.
