+++ source/extensions/common/wasm/null/BUILD
@@ -9,44 +9,45 @@ load(
 envoy_package()
 
 envoy_cc_library(
-    name = "null_plugin_hdr",
+    name = "null_vm_plugin_hdr",
     hdrs = [
-        "wasm_api_impl.h",
-    ],
-    deps = [
-        "//api/wasm/cpp:intrinsics_hdr",
+        "null_vm_plugin.h",
     ],
 )
 
 envoy_cc_library(
     name = "null_lib",
-    srcs = ["null.cc"],
-    hdrs = ["null.h"],
+    srcs = [
+        "null.cc",
+        "null_vm.cc",
+    ],
+    hdrs = [
+        "null.h",
+        "null_vm.h",
+    ],
     deps = [
-        ":null_plugin_hdr",
+        ":null_vm_plugin_hdr",
         "//external:abseil_node_hash_map",
+        "//include/envoy/registry",
         "//source/common/common:assert_lib",
-        "//source/common/common:c_smart_ptr_lib",
-        "//source/common/protobuf",
-        "//source/extensions/common/wasm:wasm_hdr",
+        "//source/extensions/common/wasm:wasm_vm_hdr",
         "//source/extensions/common/wasm:well_known_names",
-        "@envoy_api//envoy/config/wasm/v2:wasm_cc",
     ],
 )
 
-# Sample builtin Null VM/Sandbox plugin, also used in tests.
 envoy_cc_library(
-    name = "plugin",
-    srcs = [
-        "null_plugin_wrapper.cc",  # The wrapper, copy and update the name.
-        "plugin.cc",  # The plugin code.
+    name = "null_plugin_lib",
+    srcs = ["null_plugin.cc"],
+    hdrs = [
+        "null_plugin.h",
+        "wasm_api_impl.h",
     ],
-    copts = ["-DNULL_PLUGIN=1"],
     deps = [
         ":null_lib",
+        ":null_vm_plugin_hdr",
+        "//api/wasm/cpp:intrinsics_hdr",
         "//external:abseil_node_hash_map",
         "//source/common/common:assert_lib",
-        "//source/common/common:c_smart_ptr_lib",
         "//source/common/protobuf",
         "//source/extensions/common/wasm:wasm_hdr",
         "//source/extensions/common/wasm:well_known_names",
