+++ source/extensions/common/wasm/null/sample_plugin/plugin.cc
@@ -5,7 +5,7 @@
 #include "proxy_wasm_intrinsics.h"
 #else
 
-#include "extensions/common/wasm/null/null.h"
+#include "extensions/common/wasm/null/null_plugin.h"
 
 namespace Envoy {
 namespace Extensions {
