+++ source/extensions/common/wasm/v8/BUILD
@@ -20,10 +20,7 @@ envoy_cc_library(
         "//include/envoy/server:wasm_interface",
         "//include/envoy/thread_local:thread_local_interface",
         "//source/common/common:assert_lib",
-        "//source/common/common:c_smart_ptr_lib",
-        "//source/common/protobuf",
-        "//source/extensions/common/wasm:wasm_hdr",
+        "//source/extensions/common/wasm:wasm_vm_hdr",
         "//source/extensions/common/wasm:well_known_names",
-        "@envoy_api//envoy/config/wasm/v2:wasm_cc",
     ],
 )
