+++ source/extensions/common/wasm/wasm.cc
@@ -16,7 +16,6 @@
 
 #include "common/buffer/buffer_impl.h"
 #include "common/common/assert.h"
-#include "common/common/c_smart_ptr.h"
 #include "common/common/empty_string.h"
 #include "common/common/enum_to_int.h"
 #include "common/common/logger.h"
@@ -54,6 +53,27 @@ thread_local Context* current_context_ = nullptr;
 thread_local uint32_t effective_context_id_ = 0; // 0 indicates no effective context id.
 
 namespace {
+
+inline uint32_t convertWordToUint32(Word w) { return static_cast<uint32_t>(w.u64); }
+inline uint32_t convertWordToUint32(uint32_t v) { return v; }
+inline uint64_t convertWordToUint32(uint64_t v) { return v; }
+
+// Convert a function of the form Word(Word...) to one of the form uint32_t(uint32_t...).
+template <typename F, F* fn> struct ConvertFunctionWordToUint32 {
+  static void convertFunctionWordToUint32() {}
+};
+template <typename R, typename... Args, auto (*F)(Args...)->R>
+struct ConvertFunctionWordToUint32<R(Args...), F> {
+  static auto convertFunctionWordToUint32(typename ConvertWordTypeToUint32<Args>::type... args) {
+    return convertWordToUint32(F(std::forward<Args>(args)...));
+  }
+};
+template <typename... Args, auto (*F)(Args...)->void>
+struct ConvertFunctionWordToUint32<void(Args...), F> {
+  static void convertFunctionWordToUint32(typename ConvertWordTypeToUint32<Args>::type... args) {
+    F(std::forward<Args>(args)...);
+  }
+};
 
 class SharedData {
 public:
