+++ source/extensions/common/wasm/wavm/BUILD
@@ -30,10 +30,7 @@ envoy_cc_library(
     deps = [
         "//external:abseil_node_hash_map",
         "//source/common/common:assert_lib",
-        "//source/common/common:c_smart_ptr_lib",
-        "//source/common/protobuf",
-        "//source/extensions/common/wasm:wasm_hdr",
+        "//source/extensions/common/wasm:wasm_vm_hdr",
         "//source/extensions/common/wasm:well_known_names",
-        "@envoy_api//envoy/config/wasm/v2:wasm_cc",
     ],
 )
